fix: handle bfloat16 logits in dump_reference.py reranker mode
Add .float() cast before .numpy() for reranker score output,
since some models (jina-reranker-v2) default to bfloat16.

+++ tools/dump_reference.py
@@ -286,7 +286,7 @@ def dump_reranker(model_id: str, pairs: List[tuple],
             out_dict[name] = arr.astype(np.float32)
 
         # Reranker score
-        score = logits[0].detach().cpu().numpy().astype(np.float32)
+        score = logits[0].detach().cpu().float().numpy().astype(np.float32)
         out_dict[f"reranker_score_{pair_idx}"] = score
 
     return out_dict
